Feat: api 파일 이름변경 및 카테고리 리스트 api 추가
